+++ common/variables.py
